hgfhgf

+++ SpaceRTS MK2/Assets/Strategic/Empire/EmpireStats.cs
@@ -4,8 +4,13 @@ using UnityEngine;
 
 public class EmpireStats
 {
-    //Shipyards
-    public int ShipyardToolRate = 25; //Rate at which a shipyard can retool, in kT per turn.
+    ////////////////////////////////////////////////////////////////////////////////////////////////////
+    //SHIPYARDS 
+    ////////////////////////////////////////////////////////////////////////////////////////////////////
+    public float ShipyardToolRate = 25; //Rate at which a shipyard can retool, in kT per turn.
     public float ShipyardBuildRate = 1f;
 
+    
+    
+    
 }
